refactor: use double for totalPrice to support decimal values

+++ src/main/java/com/example/shop/ShoppingCart.java
@@ -2,15 +2,15 @@ package com.example.shop;
 
 public class ShoppingCart {
 
-    private int totalPrice = 0;
+    private double totalPrice = 0.0;
     private int itemCount = 0;
 
-    public void addItem(String name, int price) {
+    public void addItem(String name, double price) {
         totalPrice += price;
         itemCount++;
     }
 
-    public int getTotalPrice() {
+    public double getTotalPrice() {
         return totalPrice;
     }
